[AI battlefield] Update NVLink bandwidths to uni-directional numbers.
In the AI battlefield part, the NVLink bandwidths table quote the
bi-directional bandwidths.
As explained in the network section, it's actually deceiving - for
example for AllReduce theoretical time the relevant bandwidth is the
uni-directional one.
Also, just a couple lines below, it compares to the Ethernet bandwidth,
which are quoted as uni-directional so the comparison is not meaningful.

Switch to uni-directional bandwidths instead.